feat: add support to dotnet-installer mount auto-update with hooks
This adds the `install` hook to make sure that $SNAP_COMMON in created upon install and the `post-refresh` hook that creates the dotnet-installer-update-mounts file whenever the snap has been installed by dotnet-installer.

The dotnet-installer-update-mounts file will trigger the update mechanism on the dotnet-installer tool, which updates the mount units in the system to point to the right directory with the correct release version whenever a new content snap update is installed.

With the addition of hooks, this commit also moves the snapcraft.yaml file to the snap directory in every content snap directory.