feat: add dependency component mount units
This commit adds extra .mount units to content snaps that have
dependencies on other .NET components, e.g. the ASP.NET Core content
snap adds a .mount unit to the Microsoft.NETCore.App shared runtime
directory, which is a dependency.

The goal here is to make each content snap "standalone" in a sense that
the .NET installer won't need to install dependency content snaps to
have a fully functional installation, thus lowering the content snap
footprint on a machine.