feat: ship source-built binaries in runtime snaps
This commit changes the binaries shipped with the runtime content snaps
from the Microsoft builds to the Canonical source-built binaries by
installing the dotnet-runtime-*.0 deb package from the archive instead
of dumping a Microsoft tarball.

A side-effect of this change is that these binaries will now live in
/usr/lib/dotnet instead of /. This is a breaking change from previous
versions, which is somewhat remedied by the introduction of a dotnet
symlink in / that points to the dotnet executable in /usr/lib/dotnet.