chore: remove library-packs from .NET 9 mounts
Library packs provide NuGet packages for NativeAOT. These packages are
also provided by Microsoft through NuGet.org. Given that the library
packs directory is a top-level directory with no subdirectories by
version, it is not possible to mount it from multiple installed .NET
SDKs, such as when .NET 9 and .NET 10 SDKs are installed at once.

For that reason, I'm removing the library packs mount from .NET 9 and
not including it in .NET 10. The user will need to rely on the library
packs provided by Microsoft through NuGet.org.